Rename file css_assembler.py -> css_generator.py
